Skip explicitly setting MACOSX_DEPLOYMENT_TARGET
As we already set the `MACOSX_DEPLOYMENT_TARGET` globally before the
recipe build begins and `conda-build` respects that setting, there is no
need to set it internally in the recipe. We also target a slightly older
macOS version (10.9) to cover more users. Given this, drop the explicit
setting from the recipe so that it can use our global setting for
`MACOSX_DEPLOYMENT_TARGET`.